+++ packages/compat/src/ast-precompiler.ts
